requirement.txt
fastapi
uvicorn
python-dotenv
openai
langchain
chromadb
PyPDF2
